MOVE: Serialization related code to its own module
