+++ src/lib/kotlin/slatekit-serialization/build.gradle
@@ -22,9 +22,9 @@ apply plugin: 'kotlin'
 // ==================================================================
 // Slate Kit Component Info
 // ==================================================================
-def slatekitComponentId = 'apis'
-def slatekitComponentName = 'APIs'
-def slatekitComponentDesc = 'SlateKit APIs: Universal APIs component'
+def slatekitComponentId = 'serialization'
+def slatekitComponentName = 'Serialization'
+def slatekitComponentDesc = 'SlateKit Serialization: Reflection based serialization'
 def slatekitComponentVersion = ext.slatekit_version
 
 
@@ -51,7 +51,7 @@ task info {
     println('build.date         : ' + getBuildDate())
 }
 
-tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
+compileKotlin {
     sourceCompatibility = JavaVersion.VERSION_1_8
     targetCompatibility = JavaVersion.VERSION_1_8
 
@@ -76,25 +76,18 @@ dependencies {
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
     compile "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
     compile "com.googlecode.json-simple:json-simple:1.1"
-    compile 'com.squareup.okhttp3:okhttp:3.9.0'
     compile 'org.threeten:threetenbp:1.3.8'
 
     /* <slatekit_local>
     if( slatekitSetupViaMaven ) {
         compile "com.slatekit:slatekit-results:$slatekit_version"
         compile "com.slatekit:slatekit-common:$slatekit_version"
-        compile "com.slatekit:slatekit-context:$slatekit_version"
-        compile "com.slatekit:slatekit-meta:$slatekit_version_beta"
-        compile "com.slatekit:slatekit-policy:$slatekit_version"
-        compile "com.slatekit:slatekit-serialization:$slatekit_version"
+        compile "com.slatekit:slatekit-meta:$slatekit_version"
     } else {
         // */
         compile project(":slatekit-result")
         compile project(":slatekit-common")
-        compile project(":slatekit-context")
         compile project(":slatekit-meta")
-        compile project(":slatekit-policy")
-        compile project(":slatekit-serialization")
     // } //</slatekit_local>
 }
