+++ src/lib/kotlin/slatekit-serialization/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'slatekit-serialization'
