+++ src/lib/kotlin/slatekit-serialization/src/main/kotlin/slatekit/serialization/deserializer/DeserializeSupport.kt
@@ -1,4 +1,4 @@
-package slatekit.meta.deserializer
+package slatekit.serialization.deserializer
 
 
 interface DeserializeSupport {
