+++ src/lib/kotlin/slatekit-serialization/src/main/kotlin/slatekit/serialization/deserializer/Deserializer.kt
@@ -10,7 +10,7 @@ about: A Kotlin utility library, tool-kit and server backend.
 mantra: Simplicity above all else
 </slate_header>
  */
-package slatekit.meta.deserializer
+package slatekit.serialization.deserializer
 
 import org.json.simple.JSONObject
 import org.json.simple.parser.JSONParser
