+++ src/lib/kotlin/slatekit-serialization/src/main/kotlin/slatekit/serialization/deserializer/Deserializers.kt
@@ -1,4 +1,4 @@
-package slatekit.meta.deserializer
+package slatekit.serialization.deserializer
 
 import org.json.simple.JSONArray
 import org.json.simple.JSONObject
