+++ src/lib/kotlin/slatekit-server/build.gradle
@@ -103,6 +103,7 @@ dependencies {
         compile "com.slatekit:slatekit-apis:$slatekit_version"
         compile "com.slatekit:slatekit-jobs:$slatekit_version"
         compile "com.slatekit:slatekit-integration:$slatekit_version_beta"
+        compile "com.slatekit:slatekit-serialization:$slatekit_version"
     } else {
         // */
         compile project(":slatekit-result")
@@ -117,6 +118,7 @@ dependencies {
         compile project(":slatekit-apis")
         compile project(":slatekit-jobs")
         compile project(":slatekit-integration")
+        compile project(":slatekit-serialization")
     // } //</slatekit_local>
 }
