+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/meta/CallTests.kt
@@ -20,7 +20,7 @@ import slatekit.common.requests.InputArgs
 import slatekit.common.requests.CommonRequest
 import slatekit.common.Source
 import slatekit.meta.Reflector
-import slatekit.meta.deserializer.Deserializer
+import slatekit.serialization.deserializer.Deserializer
 import test.setup.UserApi
 
 /**
