+++ src/lib/kotlin/slatekit/src/main/kotlin/slatekit/docs/DocService.kt
@@ -33,7 +33,7 @@ class DocService(val rootdir: String, val outputDir: String, val templatePathRaw
             , Doc("Request"      ,  "slatekit-common"  , "slatekit.common.requests"           ,"slatekit.common.requests.Request"             , version, "Example_Request"          , true  , false, false, "utils", "slatekit.common.jar"  , "res"                    , "Models and abstracts a send that can represent either an HTTP send or a send from the CLI ( Command line )." )
             , Doc("SmartValues"  ,  "slatekit-common"  , "slatekit.common.smartvalues"        ,"slatekit.common.smartvalues.Email"            , version, "Example_SmartValues"      , true  , false, false, "utils", "slatekit.common.jar"  , "res"                    , "A way to store, validate and describe strongly typed and formatted strings")
             , Doc("Todo"         ,  "slatekit-common"  , "slatekit.common"                    ,"slatekit.common.Todo"                         , version, "Example_NOTE"             , true  , false, false, "utils", "slatekit.common.jar"  , "res"                    , "A programmatic approach to marking and tagging code that is strongly typed and consistent")
-            , Doc("Serialization",  "slatekit-common"  , "slatekit.common.serialization"      ,"slatekit.common.serialization.Serializer"     , version, "Example_Serialization"    , true  , false, false, "utils", "slatekit.common.jar"  , "res"                    , "Serializers for data classes to generate CSV, Props, HOCON, JSON files")
+            , Doc("Serialization",  "slatekit-common"  , "slatekit.serialization"             ,"slatekit.common.serialization.Serializer"     , version, "Example_Serialization"    , true  , false, false, "utils", "slatekit.common.jar"  , "res"                    , "Serializers for data classes to generate CSV, Props, HOCON, JSON files")
             , Doc("Templates"    ,  "slatekit-common"  , "slatekit.common.templates"          ,"slatekit.common.templates.Templates"          , version, "Example_Templates"        , true  , false, false, "utils", "slatekit.common.jar"  , "res"                    , "A micro template system for processing text with variables, useful for generating dynamic emails/messages.")
             , Doc("Validations"  ,  "slatekit-common"  , "slatekit.common.validation"         ,"slatekit.common.validation.ValidationFuncs"   , version, "Example_Validation"       , true  , false, false, "utils", "slatekit.common.jar"  , "res"                    , "A set of validation related components, simple validation checks, RegEx checks, error collection and custom validators")
             , Doc("Utils"        ,  "slatekit-common"  , "slatekit.common.utils"              ,"slatekit.common.console.ConsoleWriter"        , version, "Example_Utils"            , true  , false, false, "utils", "slatekit.common.jar"  , "res"                    , "Various utilities available in the Slate library")
